Changed package name to lowercase
