+++ src/test/java/net/playeranalytics/extension/cmi/ExtensionImplementationTest.java
@@ -20,54 +20,33 @@
     OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
     THE SOFTWARE.
 */
-package net.playeranalytics.extension.CMI;
+package net.playeranalytics.extension.cmi;
 
-import com.djrapitops.plan.extension.Caller;
 import com.djrapitops.plan.extension.DataExtension;
-import org.bukkit.Bukkit;
-
-import java.util.Optional;
+import com.djrapitops.plan.extension.extractor.ExtensionExtractor;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 /**
- * Factory for DataExtension.
+ * Test for the implementation of the new extension
  *
- * @author YannicHock
+ * @author AuroraLS3
  */
-public class CMIExtensionFactory {
+class ExtensionImplementationTest {
 
-    private boolean isAvailable() {
-        try {
-            Class.forName("com.Zrips.CMI.CMI");
-            return Bukkit.getPluginManager().isPluginEnabled("CMI");
-        } catch (ClassNotFoundException e) {
-            return false;
-        }
+    private ExtensionExtractor extractor;
+
+    @BeforeEach
+    void prepareExtractor() {
+        DataExtension extension = new CMIExtension(true);
+        extractor = new ExtensionExtractor(extension);
     }
 
-    public Optional<DataExtension> createCMIExtension() {
-        if (isAvailable()) {
-            return Optional.of(new CMIExtension());
-        }
-        return Optional.empty();
+    @Test
+    @DisplayName("API is implemented correctly")
+    void noImplementationErrors() {
+        extractor.validateAnnotations();
     }
 
-
-    public void registerCMIUpdateListeners(Caller caller) {
-        CMIEventListener.register(caller);
-    }
-
-
-    public Optional<DataExtension> createCMIEcoExtension() {
-        if (isAvailable()) {
-            CMIEcoExtension extension = new CMIEcoExtension();
-            if (CMIEcoExtension.enabled.get()) {
-                return Optional.of(extension);
-            }
-        }
-        return Optional.empty();
-    }
-
-    public void registerEcoUpdateListeners(Caller caller) {
-        CMIEcoEventListener.register(caller);
-    }
 }
